refactor: own file for pos, scale, rotate, color, opacity, anchor and area
