+++ src/components/index.ts
@@ -1,0 +1,8 @@
+export * from "./draw/circle";
+export * from "./draw/opacity";
+export * from "./misc/health";
+export * from "./physics/area";
+export * from "./transform/anchor";
+export * from "./transform/pos";
+export * from "./transform/rotate";
+export * from "./transform/scale";
