+++ src/components/physics/area.ts
@@ -1,0 +1,296 @@
+import { DEF_ANCHOR } from "../../constants";
+import { anchorPt, getInternalContext, getKaboomContext } from "../../kaboom";
+import { rgb, testPolygonPoint, vec2 } from "../../math";
+import {
+    AnchorComp,
+    AreaComp,
+    AreaCompOpt,
+    Collision,
+    EventController,
+    FixedComp,
+    GameObj,
+    MouseButton,
+    Polygon,
+    PosComp,
+    Shape,
+    Tag,
+    Vec2,
+} from "../../types";
+
+export function area(opt: AreaCompOpt = {}): AreaComp {
+    const k = getKaboomContext(this);
+    const internal = getInternalContext(k);
+    const { app } = internal;
+
+    const colliding = {};
+    const collidingThisFrame = new Set();
+
+    return {
+        id: "area",
+        collisionIgnore: opt.collisionIgnore ?? [],
+
+        add(this: GameObj<AreaComp>) {
+            if (this.area.cursor) {
+                this.onHover(() => app.setCursor(this.area.cursor));
+            }
+
+            this.onCollideUpdate((obj, col) => {
+                if (!colliding[obj.id]) {
+                    this.trigger("collide", obj, col);
+                }
+                colliding[obj.id] = col;
+                collidingThisFrame.add(obj.id);
+            });
+        },
+
+        update(this: GameObj<AreaComp>) {
+            for (const id in colliding) {
+                if (!collidingThisFrame.has(Number(id))) {
+                    this.trigger("collideEnd", colliding[id].target);
+                    delete colliding[id];
+                }
+            }
+            collidingThisFrame.clear();
+        },
+
+        drawInspect(this: GameObj<AreaComp | AnchorComp | FixedComp>) {
+            const a = this.localArea();
+
+            k.pushTransform();
+            k.pushScale(this.area.scale);
+            k.pushTranslate(this.area.offset);
+
+            const opts = {
+                outline: {
+                    width: 4 / internal.getViewportScale(),
+                    color: rgb(0, 0, 255),
+                },
+                anchor: this.anchor,
+                fill: false,
+                fixed: internal.isFixed(this),
+            };
+
+            if (a instanceof k.Rect) {
+                k.drawRect({
+                    ...opts,
+                    pos: a.pos,
+                    width: a.width,
+                    height: a.height,
+                });
+            } else if (a instanceof k.Polygon) {
+                k.drawPolygon({
+                    ...opts,
+                    pts: a.pts,
+                });
+            } else if (a instanceof k.Circle) {
+                k.drawCircle({
+                    ...opts,
+                    pos: a.center,
+                    radius: a.radius,
+                });
+            }
+
+            k.popTransform();
+        },
+
+        area: {
+            shape: opt.shape ?? null,
+            scale: opt.scale ? vec2(opt.scale) : vec2(1),
+            offset: opt.offset ?? vec2(0),
+            cursor: opt.cursor ?? null,
+        },
+
+        isClicked(): boolean {
+            return app.isMousePressed() && this.isHovering();
+        },
+
+        isHovering(this: GameObj) {
+            const mpos = internal.isFixed(this)
+                ? k.mousePos()
+                : k.toWorld(k.mousePos());
+            return this.hasPoint(mpos);
+        },
+
+        checkCollision(this: GameObj, other: GameObj<AreaComp>) {
+            return colliding[other.id] ?? null;
+        },
+
+        getCollisions() {
+            return Object.values(colliding);
+        },
+
+        // TODO: perform check instead of use cache
+        isColliding(other: GameObj<AreaComp>) {
+            return Boolean(colliding[other.id]);
+        },
+
+        isOverlapping(other) {
+            const col = colliding[other.id];
+            return col && col.hasOverlap();
+        },
+
+        onClick(
+            this: GameObj<AreaComp>,
+            f: () => void,
+            btn: MouseButton = "left",
+        ): EventController {
+            const e = app.onMousePress(btn, () => {
+                if (this.isHovering()) {
+                    f();
+                }
+            });
+            this.onDestroy(() => e.cancel());
+            return e;
+        },
+
+        onHover(this: GameObj, action: () => void): EventController {
+            let hovering = false;
+            return this.onUpdate(() => {
+                if (!hovering) {
+                    if (this.isHovering()) {
+                        hovering = true;
+                        action();
+                    }
+                } else {
+                    hovering = this.isHovering();
+                }
+            });
+        },
+
+        onHoverUpdate(this: GameObj, onHover: () => void): EventController {
+            return this.onUpdate(() => {
+                if (this.isHovering()) {
+                    onHover();
+                }
+            });
+        },
+
+        onHoverEnd(this: GameObj, action: () => void): EventController {
+            let hovering = false;
+            return this.onUpdate(() => {
+                if (hovering) {
+                    if (!this.isHovering()) {
+                        hovering = false;
+                        action();
+                    }
+                } else {
+                    hovering = this.isHovering();
+                }
+            });
+        },
+
+        onCollide(
+            this: GameObj,
+            tag: Tag | ((obj: GameObj, col?: Collision) => void),
+            cb?: (obj: GameObj, col?: Collision) => void,
+        ): EventController {
+            if (typeof tag === "function" && cb === undefined) {
+                return this.on("collide", tag);
+            } else if (typeof tag === "string") {
+                return this.onCollide((obj, col) => {
+                    if (obj.is(tag)) {
+                        cb(obj, col);
+                    }
+                });
+            }
+        },
+
+        onCollideUpdate(
+            this: GameObj<AreaComp>,
+            tag: Tag | ((obj: GameObj, col?: Collision) => void),
+            cb?: (obj: GameObj, col?: Collision) => void,
+        ): EventController {
+            if (typeof tag === "function" && cb === undefined) {
+                return this.on("collideUpdate", tag);
+            } else if (typeof tag === "string") {
+                return this.on(
+                    "collideUpdate",
+                    (obj, col) => obj.is(tag) && cb(obj, col),
+                );
+            }
+        },
+
+        onCollideEnd(
+            this: GameObj<AreaComp>,
+            tag: Tag | ((obj: GameObj) => void),
+            cb?: (obj: GameObj) => void,
+        ): EventController {
+            if (typeof tag === "function" && cb === undefined) {
+                return this.on("collideEnd", tag);
+            } else if (typeof tag === "string") {
+                return this.on(
+                    "collideEnd",
+                    (obj) => obj.is(tag) && cb(obj),
+                );
+            }
+        },
+
+        hasPoint(pt: Vec2): boolean {
+            // TODO: convert to pt to local space instead
+            return testPolygonPoint(this.worldArea(), pt);
+        },
+
+        // push an obj out of another if they're overlapped
+        resolveCollision(
+            this: GameObj<AreaComp | PosComp>,
+            obj: GameObj<AreaComp>,
+        ) {
+            const col = this.checkCollision(obj);
+            if (col && !col.resolved) {
+                this.pos = this.pos.add(col.displacement);
+                col.resolved = true;
+            }
+        },
+
+        localArea(
+            this: GameObj<AreaComp | { renderArea(): Shape }>,
+        ): Shape {
+            return this.area.shape
+                ? this.area.shape
+                : this.renderArea();
+        },
+
+        // TODO: cache
+        worldArea(this: GameObj<AreaComp | AnchorComp>): Polygon {
+            const localArea = this.localArea();
+
+            if (
+                !(localArea instanceof k.Polygon || localArea instanceof k.Rect)
+            ) {
+                throw new Error(
+                    "Only support polygon and rect shapes for now",
+                );
+            }
+
+            const transform = this.transform
+                .clone()
+                .scale(vec2(this.area.scale ?? 1))
+                .translate(this.area.offset);
+
+            if (localArea instanceof k.Rect) {
+                const offset = anchorPt(this.anchor || DEF_ANCHOR)
+                    .add(1, 1)
+                    .scale(-0.5)
+                    .scale(localArea.width, localArea.height);
+                transform.translate(offset);
+            }
+
+            return localArea.transform(transform) as Polygon;
+        },
+
+        screenArea(this: GameObj<AreaComp | FixedComp>): Polygon {
+            const area = this.worldArea();
+            if (internal.isFixed(this)) {
+                return area;
+            } else {
+                return area.transform(internal.game.cam.transform);
+            }
+        },
+
+        inspect() {
+            return `(${this.area.scale.x.toFixed(1)}, ${
+                this.area.scale.y.toFixed(1)
+            })`;
+        },
+    };
+}
